+++ projects/wickeyappstore/src/lib/ui/popover/wasprofile/wasprofile.dialog.ts
@@ -23,7 +23,7 @@ export class WasProfile {
   /**@ignore*/
   public loggedin = false;
   /** @ignore */
-  public version = '2.12.3';
+  public version = '2.12.4';
   /**@ignore*/
   constructor(
     public userService: UserService,